Refaktorerer konfigurasjon av restTemplate for saf-selvbetjening til en config-fil.
Håndterer potensiell feilmelding ved kall for å hente dokument fra saf-selvbetjening.
